A wicket puts the on-strike batter out
10 wickets resets the game.

Make the batting side fixture a function so it's really easy to reset
the game (I was trying to regenerate the battingSide state as a new
object but for some reason, the battingOrder wasn't resetting and it was
keeping the currently-in batters in when the last batter was put out).

Make `Batter` a type and stop renaming an array of Batters
unnecessarily.

App.tsx is getting quite large